Fix compile errors against the current Pico SDK

+++ firmware/usb.c
@@ -109,7 +109,7 @@ bool usb_last_report_sent(void)
 // Invoked when sent REPORT successfully to host
 // Application can use this to send the next report
 // Note: For composite reports, report[0] is report ID
-void tud_hid_report_complete_cb(uint8_t instance, uint8_t const *report, uint8_t len)
+void tud_hid_report_complete_cb(uint8_t instance, uint8_t const* report, uint16_t len)
 {
     s_has_sent_report = true;
 }
